fix: explicit int64 for NanosecondsPerYear
On a 32-bit system, an int can store values up to 2,147,483,647, which
is much smaller than 31,556,952,000,000,000 (number of nanoseconds in a
year).

+++ x/mint/test/mint_test.go
@@ -73,7 +73,7 @@ func (s *IntegrationTestSuite) TestInflationRate() {
 	require := s.Require()
 
 	type testCase struct {
-		year int
+		year int64
 		want sdktypes.Dec
 	}
 	testCases := []testCase{
